fix(backend): fix linting error
