+++ src/utils/middlewares/auth.ts
@@ -1,7 +1,12 @@
 import { NextApiResponse } from 'next';
 import { Handler } from 'next-iron-session';
 import { SessionAdminUser } from '@/pages/api/admin/login';
-import { NextIronRequest, withSession } from '../session';
+import {
+  NextIronContextHandler,
+  NextIronRequest,
+  NextIronServerSideContext,
+  withSession,
+} from '../session';
 
 /**
  * Helper function for protecting Admin only API routes.
@@ -10,7 +15,7 @@ import { NextIronRequest, withSession } from '../session';
 export function withAdminRequestSession(
   handler: Handler<NextIronRequest, NextApiResponse>,
 ) {
-  return withSession(async (req, res) => {
+  return withSession(async (req: NextIronRequest, res: NextApiResponse) => {
     const user = req.session.get('user') as SessionAdminUser;
     if (user && user.isLoggedIn && user.admin) {
       await handler(req, res);
@@ -20,5 +25,24 @@ export function withAdminRequestSession(
         message: 'Please login as an admin to access the resource',
       });
     }
+  });
+}
+
+export function withAdminAuthPage(handler: NextIronContextHandler) {
+  return withSession((ctx: NextIronServerSideContext) => {
+    const { req, res } = ctx;
+    const user = req.session.get('user') as SessionAdminUser;
+
+    if (!user || !user.isLoggedIn || !user.admin) {
+      res.setHeader('location', '/admin/login');
+      res.statusCode = 302;
+      res.end();
+      // Even if redirecting to a different location, getServerSideProps expects valid return
+      return {
+        props: {},
+      };
+    }
+
+    return handler(ctx, null);
   });
 }
